Added user state logic
